+++ src/app/datamap/dmRoute.js
@@ -4,6 +4,6 @@ module.exports = function (app) {
     // 공공데이터 API
     app.get('/app/citydata', datamap.getAllCityData);
     app.get('/app/citydata/details/:category', datamap.getCityDataByCategory);
-    app.get('/app/citydata/details/fcst', datamap.getCityDataForecast);
+
     app.get('/app/citydata/list', datamap.getCityDataSorted);
 }
